requirements.txt
absl-py==0.7.1
astor==0.8.0
blosc==1.8.1
Cython==0.29.12
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numexpr==2.6.9
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
protobuf==3.9.0
pymongo==3.8.0
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.0
tables==3.5.2
tb-nightly==1.14.0a20190603
tensorflow==2.0.0b1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
Werkzeug==0.15.4
wrapt==1.11.2
